Extract css-rule-set-to-string
